add -downloadonly command line option
This will quit SRB2 when you connect to a server, after it downloads
the files but before it actually joins. It also bypasses some checks
(e.g. server accepting joins) that are only relevant if you actually
intend on joining the server. This can be useful, for instance, if you
don't want to join a server just yet, but want to preload its files for
later.